+++ src/components/navigation/AuthNavigator.js
@@ -3,8 +3,9 @@ import { createStackNavigator } from '@react-navigation/stack';
 import NavigationConstants from 'components/navigation/NavigationConstants';
 import Login from 'components/Login';
 
+const Stack = createStackNavigator();
+
 const AuthNavigator = () => {
-  const Stack = createStackNavigator();
   const { login } = NavigationConstants;
   return (
     <Stack.Navigator>
